+++ cfripper/rules/__init__.py
@@ -23,14 +23,14 @@ from cfripper.rules.cross_account_trust import (
 from cfripper.rules.ebs_volume_has_sse import EBSVolumeHasSSERule
 from cfripper.rules.hardcoded_RDS_password import HardcodedRDSPasswordRule
 from cfripper.rules.iam_roles import IAMRolesOverprivilegedRule, IAMRoleWildcardActionOnPolicyRule
-from cfripper.rules.kms_key_wildcard_principal import KMSKeyWildcardPrincipal
+from cfripper.rules.kms_key_wildcard_principal import KMSKeyWildcardPrincipalRule
 from cfripper.rules.managed_policy_on_user import ManagedPolicyOnUserRule
 from cfripper.rules.policy_on_user import PolicyOnUserRule
 from cfripper.rules.privilege_escalation import PrivilegeEscalationRule
 from cfripper.rules.s3_bucket_policy import S3BucketPolicyPrincipalRule
 from cfripper.rules.s3_public_access import S3BucketPublicReadAclAndListStatementRule, S3BucketPublicReadWriteAclRule
 from cfripper.rules.security_group import (
-    SecurityGroupIngressOpenToWorld,
+    SecurityGroupIngressOpenToWorldRule,
     SecurityGroupMissingEgressRule,
     SecurityGroupOpenToWorldRule,
 )
@@ -44,32 +44,35 @@ from cfripper.rules.wildcard_policies import (
 from cfripper.rules.wildcard_principals import FullWildcardPrincipalRule, PartialWildcardPrincipalRule
 
 DEFAULT_RULES = {
-    "CloudFormationAuthenticationRule": CloudFormationAuthenticationRule,
-    "CrossAccountTrustRule": CrossAccountTrustRule,
-    "EBSVolumeHasSSERule": EBSVolumeHasSSERule,
-    "FullWildcardPrincipal": FullWildcardPrincipalRule,
-    "HardcodedRDSPasswordRule": HardcodedRDSPasswordRule,
-    "IAMRolesOverprivilegedRule": IAMRolesOverprivilegedRule,
-    "IAMRoleWildcardActionOnPolicyRule": IAMRoleWildcardActionOnPolicyRule,
-    "KMSKeyCrossAccountTrustRule": KMSKeyCrossAccountTrustRule,
-    "KMSKeyWildcardPrincipal": KMSKeyWildcardPrincipal,
-    "ManagedPolicyOnUserRule": ManagedPolicyOnUserRule,
-    "PartialWildcardPrincipal": PartialWildcardPrincipalRule,
-    "PolicyOnUserRule": PolicyOnUserRule,
-    "PrivilegeEscalationRule": PrivilegeEscalationRule,
-    "S3BucketPolicyPrincipalRule": S3BucketPolicyPrincipalRule,
-    "S3BucketPolicyWildcardActionRule": S3BucketPolicyWildcardActionRule,
-    "S3BucketPublicReadAclAndListStatementRule": S3BucketPublicReadAclAndListStatementRule,
-    "S3BucketPublicReadWriteAclRule": S3BucketPublicReadWriteAclRule,
-    "S3CrossAccountTrustRule": S3CrossAccountTrustRule,
-    "SecurityGroupIngressOpenToWorld": SecurityGroupIngressOpenToWorld,
-    "SecurityGroupMissingEgressRule": SecurityGroupMissingEgressRule,
-    "SecurityGroupOpenToWorldRule": SecurityGroupOpenToWorldRule,
-    "SNSTopicPolicyNotPrincipalRule": SNSTopicPolicyNotPrincipalRule,
-    "SNSTopicPolicyWildcardActionRule": SNSTopicPolicyWildcardActionRule,
-    "SQSQueuePolicyNotPrincipalRule": SQSQueuePolicyNotPrincipalRule,
-    "SQSQueuePolicyPublicRule": SQSQueuePolicyPublicRule,
-    "SQSQueuePolicyWildcardActionRule": SQSQueuePolicyWildcardActionRule,
+    rule.__name__: rule
+    for rule in (
+        CloudFormationAuthenticationRule,
+        CrossAccountTrustRule,
+        EBSVolumeHasSSERule,
+        FullWildcardPrincipalRule,
+        HardcodedRDSPasswordRule,
+        IAMRolesOverprivilegedRule,
+        IAMRoleWildcardActionOnPolicyRule,
+        KMSKeyCrossAccountTrustRule,
+        KMSKeyWildcardPrincipalRule,
+        ManagedPolicyOnUserRule,
+        PartialWildcardPrincipalRule,
+        PolicyOnUserRule,
+        PrivilegeEscalationRule,
+        S3BucketPolicyPrincipalRule,
+        S3BucketPolicyWildcardActionRule,
+        S3BucketPublicReadAclAndListStatementRule,
+        S3BucketPublicReadWriteAclRule,
+        S3CrossAccountTrustRule,
+        SecurityGroupIngressOpenToWorldRule,
+        SecurityGroupMissingEgressRule,
+        SecurityGroupOpenToWorldRule,
+        SNSTopicPolicyNotPrincipalRule,
+        SNSTopicPolicyWildcardActionRule,
+        SQSQueuePolicyNotPrincipalRule,
+        SQSQueuePolicyPublicRule,
+        SQSQueuePolicyWildcardActionRule,
+    )
 }
 
-BASE_CLASSES = {"CrossAccountCheckingRule": CrossAccountCheckingRule, "PrincipalCheckingRule": PrincipalCheckingRule}
+BASE_CLASSES = {rule.__name__: rule for rule in (CrossAccountCheckingRule, PrincipalCheckingRule)}